LayoutTests/media/video-dom-start-expected.txt

Test that when start DOM attribute is set, playback starts from that time.

TEST(video.start == 1.0) OK
TEST(video.getAttribute('start') == '1s') OK
RUN(video.load())
EVENT(canplaythrough) TEST(video.currentTime == 1.0) OK
END OF TEST
